Fixes cannot reduce access for some virtual methods
When the base virtual method is searched to merge it's visibility with
the implementing method, the code was searching for the nearest one. But
the nearest one can be a method declared with 'protected' only in the
same assembly while the base was in another assembly and declared with
'protected internal'.

To solve that the method taken into account in this case is now the
deepest virtual method we can find.