Adds ILRepack targets which automatically repacks all referenced assemblies
Typical basic usage of ILRepack is to merge all referenced assemblies. The new
ILRepack.targets achieves just that, by extending the item definition for
`@(ReferenceCopyLocalPaths)` and setting the `%(ILRepack)` metadata to `true`
for all of them by default.

Advanced users can easily modify this default behavior by creating a target that
runs before `ILRepack` and tweaking that metadata, such as:

```
  <Target Name="ILRepackAssemblies" DependsOnTargets="ResolveReferences" BeforeTargets="ILRepack" Returns="@(ReferenceCopyLocalPaths)">
    <ItemGroup>
      <ReferenceCopyLocalPaths>
        <ILRepack>false</ILRepack>
        <ILRepack Condition="$([System.String]::new('%(FileName)').StartsWith('Newtonsoft')) Or
                             $([System.String]::new('%(FileName)').StartsWith('NuGet'))">true</ILRepack>
      </ReferenceCopyLocalPaths>
    </ItemGroup>
  </Target>
```

This target runs before ILRepack and flips the default to `false` and explicitly opts in
for assemblies starting with either `Newtonsoft` or `NuGet`.

The ILRepack target uses the `GetReferenceAssemblyPaths` to retrieve the framework path, and
also uses all referenced assembly paths as references for resolution during merging. This means
that if the project could build, ILRepack should be able to load all assemblies too, since it
contains all used references too.

The merged assemblies are deleted if the `$(ILRepackDeleteMergedAssemblies)` is set to `true`
(the default). This property is added to the `.props` file so that it can be easily overriden
in the users' project if needed.

Finally, errors are captured explicitly and turned into an actual MSBuild error, to make it more
prominent and easier to diagnose.