Use the source module MVID instead of a random GUID when renaming a type
To make repacking deterministic when a type needs to be renamed, rename it to the MVID of the module it came from instead of a random GUID. The type name is guaranteed to be unique within that module.

Other than determinism it has a nice side effect of showing which of the source modules the type came from.