Fix TypeReferences in MethodDebugInformation
Previously method debug info could have TypeReferences that point to types for the using directives. It could happen that the type was declared in the assembly being merged, so it needed to be rewritten.

Without rewriting the TypeReference we had a TypeReference with a scope that wasn't in the AssemblyRef metadata table, resulting in null Scope when reading the repacked assembly.

Fixes #362