Store exported type mapping even for type forwarders to the target assembly
If we have a TypeForwardedTo and the destination assembly is being merged, we used to skip the mapping from the original exported type to the destination.

Because of that we didn't properly rewrite references to the type that was forwarded. Now we store the export mapping even if the destination is inside the target assembly, this way we can properly rewrite all references to it.

Fixes #359